requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.22.0
pandas==0.24.1
Pillow>=6.2.2
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
scipy==1.10.0
seaborn==0.9.0
six==1.12.0
torch==2.2.0
torchvision==0.23.0
